データ削除機能の実装
Fix.swiftを削除
1時間以内であれば削除できる機能を実装
