+++ custom/ContentView.swift
@@ -13,7 +13,7 @@ struct ContentView: View {
     var body: some View {
         VStack{
             DayButtonView()
-            CheckCalendarView()
+            CalendarView()
         }
     }
 }
