+++ custom/DayButtonView.swift
@@ -20,6 +20,7 @@ struct DayButtonView: View {
     
     var body: some View {
         VStack{
+            //後々リストは消す
             List {
                 ForEach(items) { item in
                     Text("Item at \(dateFormat(date: item.timestamp!))\n ElapsedDay = \(item.elapsedDay)")
@@ -47,6 +48,15 @@ struct DayButtonView: View {
                         .foregroundColor(Color.green)
                 }
 //            }
+            if let lastDate = items[items.endIndex - 1].timestamp!{
+                let calendar = Calendar(identifier: .gregorian)
+                let diff = calendar.dateComponents([.hour], from: lastDate, to: Date()).hour!
+                if diff < 1 && resetTime(date: lastDate) == resetTime(date: Date()){
+                    Button(action:deleteLastItem){
+                        Text("Delete LastData")
+                    }
+                }
+            }
             
             Spacer()
         }
@@ -68,9 +78,7 @@ struct DayButtonView: View {
             let newItem = Item(context: viewContext)
             var elapsedDay:Int64 = 0
             if items.count != 0 {
-                var startDate = items[items.startIndex].timestamp!
-                startDate = resetTime(date: startDate)
-                elapsedDay = Int64(resetTime(date: Date()).timeIntervalSince(startDate) / 86400)
+                elapsedDay = getElapsedDay()
             }
             newItem.timestamp = Date()
             newItem.elapsedDay = elapsedDay
@@ -82,7 +90,21 @@ struct DayButtonView: View {
             }
         }
     }
+    
+    private func deleteLastItem() {
+        withAnimation {
+            let context = viewContext
+            context.delete(items[items.endIndex - 1])
+            do {
+                try viewContext.save()
+            } catch {
+                let nsError = error as NSError
+                fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
+            }
+        }
+    }
 
+    //後々消す(代わりに全データ削除機能を実装する)
     private func deleteItems(offsets: IndexSet) {
         withAnimation {
             offsets.map { items[$0] }.forEach(viewContext.delete)
@@ -116,6 +138,13 @@ struct DayButtonView: View {
         components.second = 0
 
         return calendar.date(from: components)!
+    }
+    
+    func getElapsedDay() -> Int64{
+        var startDate = items[items.startIndex].timestamp!
+        startDate = resetTime(date: startDate)
+        let elapsedDay = Int64(resetTime(date: Date()).timeIntervalSince(startDate) / 86400)
+        return elapsedDay
     }
 }
